+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [06-16-2025]
 ### Added
 - TodoLists basic look
+- TodoLists view lists
 
 ### Removed
 - Homepage